create: grid for individual airline
