+++ app/javascript/components/Airline/Airline.js
@@ -1,18 +1,7 @@
 import React from 'react'
 
-const Airline = (props) => {
-  return (
-    <div className="card">
-      <div className="airline-logo">
-        <img src="{props.attributes.image_url}" alt="props.attributes.name"/>
-      </div>
-  <div className="airline-name">{props.attributes.name}</div>
-  <div className="airline-score">{props.attributes.avg_score}</div>
-      <div className="airline-link">
-        <a href={`/airlines/${props.attributes.slug}`}>View Airline</a>
-      </div>
-    </div>
-  )
+const Airline = () => {
+  return <div>This is the Airlines#show view for app.</div>
 }
 
 export default Airline
