+++ app/javascript/components/Airlines/Airlines.js
@@ -1,5 +1,6 @@
 import React, { useState, useEffect}from 'react'
 import axios from 'axios'
+import Airline from './Airline'
 
 const Airlines = () => {
   const [airlines, setAirlines] = useState([])
@@ -15,18 +16,23 @@ const Airlines = () => {
   .catch( resp => console.log(resp) )
   ;}, [airlines.length]);
 
-  const list = airlines.map(item => {
-    return (<li key={item.attributes.name}>{item.attributes.name}</li>)
+  const grid = airlines.map(item => {
+    return (
+    <Airline 
+      key={item.attributes.name}
+      attributes={item.attributes}
+    />
+    )
   })
 
-  return (
+  return ( 
     <div className="home">
       <div className="header">
         <h1>WorldFlights</h1>
         <div className="subheader">Honest, unbiased airline reviews.</div>
       </div>
       <div className="grid">
-        <ul>{list}</ul>
+        {grid}
       </div>
     </div>
   )
