+++ app/serializers/airline_serializer.rb
@@ -1,5 +1,5 @@
 class AirlineSerializer
   include FastJsonapi::ObjectSerializer
-  attributes :name, :image_url, :slug
+  attributes :name, :image_url, :slug, :avg_score
   has_many :reviews
 end
